% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_action.R
\name{action}
\alias{action}
\title{Choose taken action according to a given state}
\usage{
action(policy, state, limit_FUN = function(x) x, ...)
}
\arguments{
\item{policy}{nn}

\item{state}{data.frame or list (usually from the 'state' function)}

\item{limit_FUN}{function to clip actions (min and max)}

\item{...}{optional : additional parameters for limit_FUN}
}
\value{
data.frame of the actions.
}
\description{
Choose taken action according to a given state
}
\examples{
action(neuralnetwork(x+y~a+b, hidden=2),
                 state(neuralnetwork(x+y~a+b, hidden=2)$formula,
                 data.frame(x=2, b=5, a=8, y=4)),
                 function(x) pmin(pmax(x, c(-1, 0)), c(1, 1)))
}
