% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backprop.R
\name{backprop}
\alias{backprop}
\title{Backpropagate data on a neural network}
\usage{
backprop(nn, newdata, threshold = 1e-04, stepmax = 10000, step_size = 0.5)
}
\arguments{
\item{nn}{neural network}

\item{newdata}{data.frame of the inputs and the outputs}

\item{threshold}{numeric : maximum gradient of all the weights to stop backpropagation}

\item{stepmax}{integer : max number of iterations of backpropagation}

\item{step_size}{numeric : multiplier of the gradient.}
}
\value{
list of neural network (class : nn) and the mean Loss
}
\description{
Backpropagate data on a neural network
}
