% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nn.R
\name{color_gradient.nn}
\alias{color_gradient.nn}
\title{Create hexadecimal color between red and green for a value [0, 1]}
\usage{
color_gradient.nn(u)
}
\arguments{
\item{u}{numeric vector}
}
\value{
hexadecimal character color.
}
\description{
Create hexadecimal color between red and green for a value [0, 1]
}
\examples{
color_gradient.nn(c(0, 0.25, 0.5, 0.75, 1))
}
