% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\title{Models parameters}
\usage{
config(formula = NULL, k = NULL)
}
\arguments{
\item{formula}{formula parameter for eg. linear models including lm, rlm, read more: \link[stats]{lm}}

\item{k}{number of neighbors considered from knn models, read more: \link[class]{knn}}
}
\value{
configuration list contains models parameters different than defaults
}
\description{
The \code{config} function sets additional models parameters
}
\examples{
\donttest{

config(formula = "Status ~ Value")

}

}
