% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backprop.R
\name{dLReLU}
\alias{dLReLU}
\title{Leaky Rectified Linear Unit activation function}
\usage{
dLReLU(x, a = 0.1)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
derivative of Leaky ReLUed arguements.
}
\description{
Leaky Rectified Linear Unit activation function
}
\examples{
dLReLU(c(3, 5, -2, 0))
dLReLU(matrix(-2:3, 2))
}
