% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralnetwork.R
\name{dReLU}
\alias{dReLU}
\title{Rectified Linear Unit activation function}
\usage{
dReLU(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
derivative of ReLUed arguements.
}
\description{
Rectified Linear Unit activation function
}
\examples{
dReLU(c(3, 5, -2, 0))
dReLU(matrix(-2:3, 2))
}
