% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backprop.R
\name{errors}
\alias{errors}
\title{calculate errors in neural network for derivatives.
Errors are derivatives of the cost function with regards to the z (XW+b).}
\usage{
errors(
  nn,
  acti,
  inputs_indiv,
  target_indiv = NULL,
  Loss_fun = TRUE,
  policy_linear_output = TRUE
)
}
\arguments{
\item{nn}{neural network object.}

\item{inputs_indiv}{data.frame of one observations with every input.}

\item{target_indiv}{data.frame of one observations with every real target.}

\item{Loss_fun}{logical : if the function to maximize/minimize is a Loss function.
if FALSE, it is considered to be a linear target.}

\item{policy_linear_output}{logical : if the policy output in DDPG is linear.}
}
\value{
list of matrices of errors for derivatives.
}
\description{
calculate errors in neural network for derivatives.
Errors are derivatives of the cost function with regards to the z (XW+b).
}
