% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backprop.R
\name{gradients}
\alias{gradients}
\title{calculate gradients in neural network.}
\usage{
gradients(nn, error, acti)
}
\arguments{
\item{nn}{neural network object.}

\item{error}{list of errors.}
}
\value{
list of matrices of errors for derivatives.
}
\description{
calculate gradients in neural network.
}
