% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backprop.R
\name{LReLU}
\alias{LReLU}
\title{Leaky Rectified Linear Unit activation derivative function}
\usage{
LReLU(x, a = 0.1)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
Leaky ReLUed arguments.
}
\description{
Leaky Rectified Linear Unit activation derivative function
}
\examples{
LReLU(c(3, 5, -2, 0))
LReLU(matrix(-2:3, 2))
}
