% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.nn.R
\name{plot.nn}
\alias{plot.nn}
\title{Plot a neural network}
\usage{
\method{plot}{nn}(nn, inputs_indiv = NULL)
}
\arguments{
\item{nn}{neural network}

\item{inputs_indiv}{optional : data.frame of one observations with every input to show activation in neurons.}
}
\value{
plot
}
\description{
Plot a neural network
}
