% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.nn.R
\name{predict.nn}
\alias{predict.nn}
\title{Predict outputs of neural network}
\usage{
\method{predict}{nn}(nn, newdata)
}
\arguments{
\item{nn}{nn object}

\item{newdata}{data.frame with at least the inputs variables as columns.}
}
\value{
matrix of the output neurons.
}
\description{
Predict outputs of neural network
}
