% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_nn.R
\name{predict_nn}
\alias{predict_nn}
\title{Predict outputs of neural network}
\usage{
predict_nn(nn, newdata)
}
\arguments{
\item{nn}{nn object}

\item{newdata}{data.frame with at least the inputs variables as columns.}
}
\value{
matrix of the output neurons.
}
\description{
Predict outputs of neural network
}
