% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backprop.R
\name{ReLU}
\alias{ReLU}
\title{Rectified Linear Unit activation derivative function}
\usage{
ReLU(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
ReLUed arguments.
}
\description{
Rectified Linear Unit activation derivative function
}
\examples{
ReLU(c(3, 5, -2, 0))
ReLU(matrix(-2:3, 2))
}
