% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDPG.R
\name{see_DDPG}
\alias{see_DDPG}
\title{See convergence of weights}
\usage{
see_DDPG(tracking)
}
\arguments{
\item{tracking}{matrix of every weights and biaises.}
}
\value{
ggplot.
}
\description{
See convergence of weights
}
\examples{
see_DDPG(DDPG(<...>)$tracking)
}
