% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDPG.R
\name{see_variables_impact}
\alias{see_variables_impact}
\title{See effect of every inputs on actions or rewards}
\usage{
see_variables_impact(nn)
}
\arguments{
\item{policy_nn}{policy neural network.}

\item{critic_nn}{critic neural network.}

\item{policy}{logical : if policy is passed.}
}
\value{
ggplot2.
}
\description{
See effect of every inputs on actions or rewards
}
