% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralnetwork.R
\name{sig}
\alias{sig}
\title{Sigmoid function}
\usage{
sig(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
sigmoized arguments.
}
\description{
Sigmoid function
}
\examples{
sig(c(3, 5, -2, 0))
sig(matrix(1:6, 2))
}
