% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_action.R
\name{state}
\alias{state}
\title{Create state of an object}
\usage{
state(formula, object)
}
\arguments{
\item{formula}{formula of the parameters that counts as state. Only right part with ~ is needed.}

\item{object}{data.frame or list of the object}
}
\value{
data.frame of the variable used for state.
}
\description{
Create state of an object
}
\examples{
state(formula=neuralnetwork(x+y~a+b, hidden=2)$formula, object=list(x=3, y=5, a=3, b=8))
}
