% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats}
\alias{stats}
\title{Models statistics}
\usage{
stats(modelA, modelB = NULL)
}
\arguments{
\item{modelA}{Model generated by \link[ai]{model} function}

\item{modelB}{Model generated by \link[ai]{model} function}
}
\value{
modified model list contains additional statistics
}
\description{
The \code{stats} function calculates models statistics. Read more \link[Metrics]{auc}
}
\examples{
\donttest{

model_data <- data.frame(a = c(1,2,3,4,5,6),
                         b = c(1,2,3,4,5,6),
                         s = c(1,2,3,4,5,6))

model_data <- prodata(model_data, status_colname = "s")

config <- config(formula = "a ~ b + s")

model <- model(model_data, config)

stats(model)

}

}
