% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats_compare_models}
\alias{stats_compare_models}
\title{stats_compare_models()}
\usage{
stats_compare_models(modelA, modelB)
}
\arguments{
\item{modelA}{modelA}

\item{modelB}{modelB}
}
\value{
data.frame contains comparison of both models statistics
}
\description{
stats_compare_models()
}
