% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backprop.R
\name{activations}
\alias{activations}
\title{calculate activations in neural network.}
\usage{
activations(nn, inputs)
}
\arguments{
\item{nn}{neural network object.}

\item{inputs}{matrix of observations with every input.}
}
\value{
list of matrices of activations.
}
\description{
calculate activations in neural network.
}
